requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.6.0
cffi==1.15.1
Django==4.1.9
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
importlib-metadata==6.6.0
Markdown==3.4.3
mysqlclient==2.1.1
pycparser==2.21
PyJWT==2.7.0
pytz==2023.3
sqlparse==0.4.4
zipp==3.15.0
